+++ src/org/labkey/test/util/JUnitHeader.java
@@ -48,9 +48,8 @@ public class JUnitHeader extends BaseWebDriverTest
         PipelineToolsHelper pipelineToolsHelper = new PipelineToolsHelper(this);
         pipelineToolsHelper.setToolsDirToTestDefault(); // Point to extra tools if present (currently only sequeneanalysis tools)
 
-        try{
+        if (_containerHelper.doesContainerExist("Shared/_junit"))
             _containerHelper.deleteFolder("Shared", "_junit");
-        }catch(Throwable e){/*ignore*/}
 
         startSystemMaintenance();
